requirements.txt
Pillow~=10.1.0
pyzbar~=0.1.9